fix lint

+++ lib/database/src/data_storage/mmap_data_storage.rs
@@ -60,7 +60,7 @@ impl MmapDataStorage {
 
     fn check_storage_overflow<V: Deref<Target = [u8]>>(&self, row: &RowToWrite<V>) -> bool {
         let row_size = self.formatter.row_size(row);
-        (row_size + self.write_offset as usize) > self.options.max_data_file_size
+        (row_size + self.write_offset) > self.options.max_data_file_size
     }
 
     fn as_mut_slice(&mut self) -> &mut [u8] {
